move utils to file within directory
