+++ src/components/Course/index.tsx
@@ -9,7 +9,7 @@ import {
   faTrash,
 } from '@fortawesome/free-solid-svg-icons';
 
-import { classes, getContentClassName } from '../../utils';
+import { classes, getContentClassName } from '../../utils/misc';
 import { ActionRow, Instructor, Palette, Prerequisite } from '..';
 import { TermContext } from '../../contexts';
 import { Course as CourseBean, Section } from '../../data/beans';
