+++ src/components/CourseAdd/index.tsx
@@ -11,7 +11,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 
 import { Course, CourseFilter } from '..';
-import { classes, getRandomColor } from '../../utils';
+import { classes, getRandomColor } from '../../utils/misc';
 import { ASYNC_DELIVERY_MODE, CAMPUSES, DELIVERY_MODES } from '../../constants';
 import { TermContext } from '../../contexts';
 import { Course as CourseBean, Section } from '../../data/beans';
