+++ src/components/CourseFilter/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { faAngleDown, faAngleUp } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { classes, humanizeArrayReact } from '../../utils';
+import { classes, humanizeArrayReact } from '../../utils/misc';
 
 import './stylesheet.scss';
