+++ src/components/Prerequisite/index.tsx
@@ -6,7 +6,7 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 
 import { TermContext } from '../../contexts';
-import { classes, serializePrereqs } from '../../utils';
+import { classes, serializePrereqs } from '../../utils/misc';
 import { ActionRow } from '..';
 import { Course } from '../../data/beans';
 import {
