+++ src/components/Scheduler/index.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo, useState } from 'react';
 
-import { classes } from '../../utils';
+import { classes } from '../../utils/misc';
 import { Button, Calendar, CombinationContainer, CourseContainer } from '..';
 import { OverlayCrnsContext, OverlayCrnsContextValue } from '../../contexts';
 import { useMobile } from '../../hooks';
