+++ src/components/Spinner/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { classes } from '../../utils';
+import { classes } from '../../utils/misc';
 
 import './stylesheet.scss';
